Split generic primitive contract tests
