+++ package.json
@@ -65,7 +65,11 @@
     "test:runtime-recipe-resolver": "tsx tests/runtime-recipe-resolver.test.ts",
     "test:browser-runner-template": "tsx tests/browser-runner-template.test.ts",
     "test:editor-actions": "tsx tests/editor-actions.test.ts",
-    "test:generic-primitives": "tsx tests/generic-primitives.test.ts",
+    "test:artifact-path-primitives": "tsx tests/artifact-path-primitives.test.ts",
+    "test:browser-callback-materialization-contracts": "tsx tests/browser-callback-materialization-contracts.test.ts",
+    "test:source-package-compiler-primitives": "tsx tests/source-package-compiler-primitives.test.ts",
+    "test:bench-command-step-behavior": "tsx tests/bench-command-step-behavior.test.ts",
+    "test:generic-primitives": "npm run test:artifact-path-primitives && npm run test:browser-callback-materialization-contracts && npm run test:source-package-compiler-primitives && npm run test:bench-command-step-behavior",
     "test:browser-artifact-session": "tsx tests/browser-artifact-session.test.ts",
     "test:browser-provider-permissions": "tsx tests/browser-provider-permissions.test.ts",
     "test:browser-provider-bridge-inheritance": "tsx tests/browser-provider-bridge-inheritance.test.ts",
